last_resort/last_resort_requirements.txt
fpdf2==2.8.1
psutil==6.1.0
PyYAML==6.0.1